kpm: Create npm init json and file to update it with user inputs

+++ bin/scaffold/scaffold-module.js
@@ -1,11 +1,13 @@
 // Community
+const fs = require('fs');
 const fse = require('fs-extra');
 const path = require('path');
 
 // Internal
 const { whereAmI, isDrupalInstall, getModulesFolderPath } = require('../utils/path-utils');
 const { renameBaseFiles } = require('./build/rename-base-files');
-const {updateModuleFile} = require("./build/update-module-file");
+const { updateModuleFile } = require('./build/update-module-file');
+const { updateRootScaffoldFile } = require('./build/update-scaffold-file');
 
 /**
  * @description This creates a module based on the users inputs
@@ -25,9 +27,13 @@ const scaffoldModule = function(answers) {
   const modulePath = `${customPath}/${machineName}`;
 
   try {
+    if (fs.existsSync(modulePath)) {
+      console.log('There is already a module of that name. Please use another name.');
+
+      process.exit(0);
+    }
+
     // Copy our files over to the modules folder
-    // @todo We need to let the user know they have already copied over the files if they use the same name
-    // @todo We need to update this path for development
     fse.copySync(`${path.join(__dirname + '../../../scaffolding')}`, modulePath, {overwrite: false});
 
     // Rename the YML and install file based on user input
@@ -41,6 +47,14 @@ const scaffoldModule = function(answers) {
       moduleAdminName,
       moduleDescriptionName,
     });
+
+    // Update the package.json
+    updateRootScaffoldFile(
+      modulePath,
+      moduleAdminName,
+      moduleDescriptionName,
+      'package.json'
+    );
 
   } catch (err) {
